Rename the auto/export pkg to auto/pipeline
