+++ Dockerfile.dockerignore
@@ -8,6 +8,7 @@
 
 # directories
 !/cli
-!/sdk
 !/internal/include
 !/internal/pkg
+!/pipeline
+!/sdk
