+++ pipeline/handler_test.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package export_test
+package pipeline_test
 
 import (
 	"testing"
