+++ pipeline/otelsdk/handler.go
@@ -1,9 +1,9 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-// Package otelsdk provides an implementation of [export.Handler] that uses the
-// default OpenTelemetry Go SDK to process and export the telemetry generated
-// by auto-instrumentation.
+// Package otelsdk provides an implementation of [pipeline.Handler] that uses
+// the default OpenTelemetry Go SDK to process and export the telemetry
+// generated by auto-instrumentation.
 package otelsdk
 
 import (
@@ -19,18 +19,18 @@ import (
 	sdk "go.opentelemetry.io/otel/sdk/trace"
 	"go.opentelemetry.io/otel/trace"
 
-	"go.opentelemetry.io/auto/export"
+	"go.opentelemetry.io/auto/pipeline"
 )
 
-// NewHandler returns a new configured [export.Handler] that uses the
+// NewHandler returns a new configured [pipeline.Handler] that uses the
 // OpenTelemetry SDK (go.opentelemetry.io/otel/sdk) to process and export
 // telemetry generated by auto-instrumentation.
-func NewHandler(ctx context.Context, options ...Option) (*export.Handler, error) {
+func NewHandler(ctx context.Context, options ...Option) (*pipeline.Handler, error) {
 	th, err := NewTraceHandler(ctx, options...)
 	if err != nil {
 		return nil, err
 	}
-	return &export.Handler{TraceHandler: th}, nil
+	return &pipeline.Handler{TraceHandler: th}, nil
 }
 
 // TraceHandler handles telemetry produced by auto-instrumentation by processing
@@ -42,7 +42,7 @@ type TraceHandler struct {
 	stopped atomic.Bool
 }
 
-var _ export.TraceHandler = (*TraceHandler)(nil)
+var _ pipeline.TraceHandler = (*TraceHandler)(nil)
 
 // NewTraceHandler returns a new configured TraceHandler that uses the
 // OpenTelemetry SDK (go.opentelemetry.io/otel/sdk) to process and export
